requirements.txt
pandas
numpy
pydicom
matplotlib
Pillow
scikit-image
seaborn
torch
torchvision
torchmetrics
gradio
opencv-python
